+++ R/get_residual_cor.R
@@ -8,15 +8,22 @@
 #' @name get_residual_cor
 #' @aliases get_residual_cor
 #' @title Calculate the residual correlation matrix from a latent variable model (LVM)
-#' @description This function use coefficients \eqn{(\lambda_{jl} with j=1,\dots,n_species and l=1,\dots,n_latent)}{(\lambda_jl with j=1,...,n_species and l=1,...,n_latent)}, corresponding to latent variables fitted using \code{jSDM} package, to calculate the variance-covariance matrix which controls correlation between species.
+#' @description This function use coefficients \eqn{(\lambda_{jl} with j=1,\dots,n_{species} and l=1,\dots,n_{latent})}{(\lambda_jl with j=1,...,n_species and l=1,...,n_latent)}, corresponding to latent variables fitted using \code{jSDM} package, to calculate the variance-covariance matrix which controls correlation between species.
 #' @param mod An object of class \code{"jSDM"}
+#' @param prob A numeric scalar in the interval \eqn{(0,1)} giving the target probability coverage of the intervals, by which to determine whether the correlations are "significant". Defaults to 0.95.
 #' @return results A list including : 
 #' \item{cov.mean}{Average over the MCMC samples of the variance-covariance matrix.} 
 #' \item{cov.median}{Median over the MCMC samples of the variance-covariance matrix.}
+#' \item{cov.lower}{A \eqn{n_{species} \times n_{species}}{n_species x n_species} matrix containing the lower limits of the  (\eqn{100 \times prob \%}{100 x prob \%}) HPD interval of variance-covariance matrices over the MCMC samples.}
+#' \item{cov.upper}{A \eqn{n_{species} \times n_{species}}{n_species x n_species} matrix containing the upper limits of the  (\eqn{100 \times prob \%}{100 x prob \%}) HPD interval of variance-covariance matrices over the MCMC samples.}
+#' \item{sig.cov}{A \eqn{n_{species} \times n_{species}}{n_species x n_species} matrix containing the value 0 corresponding to non-significant co-variances and the value 1 corresponding to the “significant" co-variances whose (\eqn{100 \times prob \%}{100 x prob \%}) HPD interval over the MCMC samples does not contain zero.}
 #' \item{cor.mean}{Average over the MCMC samples of the residual correlation matrix.}
 #' \item{cor.median}{Median over the MCMC samples of the residual correlation matrix.}
-#' 
-#' @details  After fitting the jSDM with latent variables, the \bold{fullspecies residual correlation matrix} : \eqn{R=(R_ij) avec i=1,\ldots, nspecies et j=1,\ldots, nspecies}{R=(R_ij) avec i=1,..., nspecies et j=1,..., nspecies} can be derived from the covariance in the latent variables such as : 
+#' \item{cor.lower}{A \eqn{n_{species} \times n_{species}}{n_species x n_species} matrix containing the lower limits of the  (\eqn{100 \times prob \%}{100 x prob \%}) HPD interval of correlation matrices over the MCMC samples.}
+#' \item{cor.upper}{A \eqn{n_{species} \times n_{species}}{n_species x n_species} matrix containing the upper limits of the  (\eqn{100 \times prob \%}{100 x prob \%}) HPD interval of correlation matrices over the MCMC samples.}
+#' \item{sig.cor}{A \eqn{n_{species} \times n_{species}}{n_species x n_species} matrix containing the value 0 corresponding to non-significant correlations and the value 1 corresponding to the “significant" correlations whose (\eqn{100 \times prob \%}{100 x prob \%})  HPD interval over the MCMC samples does not contain zero.}
+#'
+#' @details  After fitting the jSDM with latent variables, the \bold{fullspecies residual correlation matrix} : \eqn{R=(R_ij) avec i=1,\ldots, n_{species} et j=1,\ldots, n_{species}}{R=(R_ij) avec i=1,..., n_species et j=1,..., n_species} can be derived from the covariance in the latent variables such as : 
 #' \tabular{lll}{
 #' \eqn{\Sigma_{ij}}{Sigma_ij} \tab \eqn{= \lambda_i .\lambda_j' + 1}{= \lambda_i . \lambda_j' + 1} \tab if i=j \cr
 #'          \tab \eqn{= \lambda_i .\lambda_j'}{= \lambda_i . \lambda_j'} \tab else, \cr}
@@ -66,24 +73,42 @@
 #'                              # Various
 #'                              seed=1234, verbose=1)
 #' # Calcul of residual correlation between species 
-#'  result <- get_residual_cor(mod)
-#'  result$cov.mean
-#'  result$cor.mean
+#' result <- get_residual_cor(mod, prob=0.95)
+#' # Residual variance-covariance matrix
+#' result$cov.mean
+#' ## All non-significant co-variances are set to zero.
+#' result$cov.mean * result$sig.cov
+#' # Residual correlation matrix
+#' result$cor.mean
+#' ## All non-significant correlations are set to zero.
+#' result$cor.mean * result$sig.cor
 #' @keywords stats::cov2cor
 #' @importFrom stats cov2cor
+#' @importFrom coda HPDinterval
 #' @export
 
 
 # Calculate the residual correlation matrix from a LVM
-get_residual_cor <- function(mod) {
+get_residual_cor <- function(mod, prob=0.95) {
   #== Check
-  if (!class(mod)=="jSDM"){
+  if(!class(mod)=="jSDM"){
     stop("Please provide an object of class jSDM in", calling.function(), call.=FALSE)
   }
-  if (mod$model_spec$n_latent==0) {
+  if(mod$model_spec$n_latent==0) {
     cat("Error: The jSDM class object provided is not a latent variable model (LVM).\n
-        The lambdas parameters needed to compute the residual correlation matrix have not been estimated. \n")
+        The factor loadings needed to compute the residual correlation matrix have not been estimated. \n")
     stop("Please fit a LVM on your data and call ", calling.function(), " again.",
+         call.=FALSE)
+  }
+  if(mod$model_spec$n_latent==1) {
+    cat("Error: Residual correlation matrix is reliably modeled only with two or more latent variables. \n")
+    stop("Please fit a LVM with n_latent > 1 on your data and call ", calling.function(), " again.",
+         call.=FALSE)
+  }
+  
+  if(prob>1 || prob<0) {
+    cat("Error: The target probability coverage of the intervals must be in the interval ]0,1[. \n")
+    stop("Please specify a probability for prob and call ", calling.function(), " again.",
          call.=FALSE)
   }
   if(!is.null(mod$model_spec$presence_data)){
@@ -106,13 +131,49 @@ get_residual_cor <- function(mod) {
     Tau.cor.mat <- cov2cor(Tau.mat)
     Tau.cor.arr[t,] <- as.vector(Tau.cor.mat) 
   }
-  
+  # Lower/upper limits of HPD interval over the MCMC samples
+  cov.lower <- cov.upper <- cor.lower <- cor.upper <-  matrix(NA, n.species, n.species)
+  sig.Tau.cor <- sig.Tau.mat <- matrix(NA, n.species, n.species)
+  for(j in 1:n.species){
+    for(jprim in 1:n.species){
+      ## Residual variance-covariance matrices
+      get.hpd.covs <- coda::HPDinterval(coda::as.mcmc(Tau.arr[,(j-1)*n.species + jprim]),
+                                        prob = prob)
+      cov.lower[j, jprim] <- get.hpd.covs[1]
+      cov.upper[j, jprim] <- get.hpd.covs[2]
+      ## Significant values whose HPD interval does not contain zero
+      sig.Tau.mat[j, jprim] <- ifelse((0 > get.hpd.covs[1] & 0 < get.hpd.covs[2]), 0, 1)
+      ## Residual correlations matrices
+      get.hpd.cors <- coda::HPDinterval(coda::as.mcmc(Tau.cor.arr[,(j-1)*n.species + jprim]),
+                                        prob = prob)
+      cor.lower[j, jprim] <- get.hpd.cors[1]
+      cor.upper[j, jprim] <- get.hpd.cors[1]
+      ## Significant values whose HPD interval does not contain zero
+      sig.Tau.cor[j, jprim] <-  ifelse((0 > get.hpd.cors[1] & 0 < get.hpd.cors[2]), 0, 1)
+    }
+  }
   ## Average/Median over the MCMC samples
-  Tau.mat.mean <- sig.Tau.mat.mean <- matrix(apply(Tau.arr,2,mean),n.species,byrow=F)
-  Tau.mat.median <- sig.Tau.mat.median <- matrix(apply(Tau.arr,2,median),n.species,byrow=F)
-  Tau.cor.mean <- sig.Tau.cor.mean <- matrix(apply(Tau.cor.arr,2,mean),n.species,byrow=F)
-  Tau.cor.median <- sig.Tau.cor.median <- matrix(apply(Tau.cor.arr,2,median),n.species,byrow=F)
-  results <- list(cov.mean = Tau.mat.mean, cov.median = Tau.mat.median, cor.mean = Tau.cor.mean, cor.median = Tau.cor.median)
+  Tau.mat.mean <-  matrix(apply(Tau.arr,2,mean),n.species,byrow=F)
+  Tau.mat.median <-  matrix(apply(Tau.arr,2,median),n.species,byrow=F)
+  Tau.cor.mean <- matrix(apply(Tau.cor.arr,2,mean),n.species,byrow=F)
+  Tau.cor.median <- matrix(apply(Tau.cor.arr,2,median),n.species,byrow=F)
+  # Species names in results 
+  if(!is.null(mod$model_spec$presence_data)){
+    names.sp <- colnames(mod$model_spec$presence_data)
+  }
+  if(!is.null(mod$model_spec$count_data)){
+    names.sp <-  colnames(mod$model_spec$count_data)
+  }
+  if(!is.null(mod$model_spec$data)){
+    names.sp <- unique(mod$model_spec$data$species)
+  }
+  dimnames(Tau.mat.mean) <- dimnames(Tau.mat.median) <- dimnames(cov.lower) <- dimnames(cov.upper) <- dimnames(sig.Tau.mat) <- list(names.sp, names.sp)
+  dimnames(Tau.cor.mean) <- dimnames(Tau.cor.median) <- dimnames(cor.lower) <- dimnames(cor.upper) <- dimnames(sig.Tau.cor) <- list(names.sp, names.sp)
+  # Results 
+  results <- list(cov.mean = Tau.mat.mean, cov.median = Tau.mat.median,
+                  cov.lower=cov.lower, cov.upper=cov.upper, cov.sig=sig.Tau.mat,
+                  cor.mean = Tau.cor.mean, cor.median = Tau.cor.median,
+                  cor.lower=cor.lower, cor.upper=cor.upper, cor.sig=sig.Tau.cor)
   return(results)
 }